+++ libs/tendermint/types/block.go
@@ -3,7 +3,6 @@ package types
 import (
 	"bytes"
 	"fmt"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -37,33 +36,6 @@ const (
 	MaxAminoOverheadForBlock int64 = 11
 )
 
-// GetStartBlockHeight() is the block height from which the chain starts
-var (
-	startBlockHeightStr       = "0"
-	startBlockHeight    int64 = 0
-	once                sync.Once
-)
-
-func initStartBlockHeight() {
-	once.Do(func() {
-		var err error
-		if len(startBlockHeightStr) == 0 {
-			startBlockHeightStr = "0"
-		}
-		startBlockHeight, err = strconv.ParseInt(startBlockHeightStr, 10, 64)
-		if err != nil {
-			panic(err)
-		}
-	})
-}
-
-func init() {
-	initStartBlockHeight()
-}
-
-func GetStartBlockHeight() int64 {
-	return startBlockHeight
-}
 
 // Block defines the atomic unit of a Tendermint blockchain.
 type Block struct {
@@ -1200,12 +1172,3 @@ func BlockIDFromProto(bID *tmproto.BlockID) (*BlockID, error) {
 	return blockID, blockID.ValidateBasic()
 }
 
-// 2322600 is mainnet GenesisHeight
-func IsMainNet() bool {
-	return startBlockHeightStr == "2322600"
-}
-
-// 1121818 is testnet GenesisHeight
-func IsTestNet() bool {
-	return startBlockHeightStr == "1121818"
-}
